fix PSO variants
