+++ src/algorithms/__init__.py
@@ -1,9 +1,1 @@
-from .clpso import CLPSO
-from .cso import CSO
-from .dms_pso_el import DMSPSOEL
-from .fips import FIPS
-from .fs_pso import FSPSO
-from .pso import PSO
-from .sl_pso_gs import SLPSOGS
-from .sl_pso_us import SLPSOUS
-from .swmmpso import SWMMPSO
+from .pso_variants import *
