+++ src/algorithms/pso_variants/__init__.py
@@ -1,0 +1,7 @@
+from .clpso import CLPSO
+from .cso import CSO
+from .dms_pso_el import DMSPSOEL
+from .fs_pso import FSPSO
+from .pso import PSO
+from .sl_pso_gs import SLPSOGS
+from .sl_pso_us import SLPSOUS
